Renamed general to set.
